Move kxBans into its own file
